moved getColorLumiance to its own util file. removed unused package. removed Padding from inside Case Study
